backend/requirements.txt
fastapi
uvicorn
python-dotenv
paho-mqtt
requests
websockets
